Corrigir erro de tipo token dashboard

+++ server.js
@@ -1204,8 +1204,10 @@ app.get('/api/eventos', async (req, res) => {
         END as status_envio,
         'tokens' as source_table
       FROM tokens t
-      LEFT JOIN tracking_data td ON t.telegram_id = td.telegram_id
-      LEFT JOIN payload_tracking pt ON t.telegram_id = pt.telegram_id
+      -- ✅ CORREÇÃO: Cast de t.telegram_id (TEXT) para BIGINT para compatibilidade 
+      -- com td.telegram_id e pt.telegram_id que são BIGINT
+      LEFT JOIN tracking_data td ON t.telegram_id::bigint = td.telegram_id
+      LEFT JOIN payload_tracking pt ON t.telegram_id::bigint = pt.telegram_id
       LEFT JOIN payloads p ON t.token = p.payload_id
       WHERE (t.pixel_sent = true OR t.capi_sent = true OR t.cron_sent = true)
       
@@ -1295,8 +1297,9 @@ app.get('/api/eventos', async (req, res) => {
           t.valor,
           COALESCE(td.utm_source, pt.utm_source, p.utm_source) as utm_source
         FROM tokens t
-        LEFT JOIN tracking_data td ON t.telegram_id = td.telegram_id
-        LEFT JOIN payload_tracking pt ON t.telegram_id = pt.telegram_id
+        -- ✅ CORREÇÃO: Cast de t.telegram_id (TEXT) para BIGINT para compatibilidade
+        LEFT JOIN tracking_data td ON t.telegram_id::bigint = td.telegram_id
+        LEFT JOIN payload_tracking pt ON t.telegram_id::bigint = pt.telegram_id
         LEFT JOIN payloads p ON t.token = p.payload_id
         WHERE (t.pixel_sent = true OR t.capi_sent = true OR t.cron_sent = true)
